Fixed a bug where a message sent with a multiple of 56 bytes would fail. Corrected input length to be size_t to fix issue with bytes > 255. Cleaned update code to make more readable.

+++ rsa/sha.cpp
@@ -60,7 +60,7 @@ char *SHA_MODE_NAME(SHA_MODE mode)
     }
     return "";
 }
-int sha_update(uint8_t *msg, uint8_t byMsg_len, SHA_Context *ctx)
+int sha_update(uint8_t *msg, size_t byMsg_len, SHA_Context *ctx)
 {
     switch(ctx->mode)
     {
